ocr-service/pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
description = "Card image OCR microservice (EasyOCR); used by vc-trading-cards backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "easyocr>=1.7.0",
    "Pillow>=10.0.0",
    "numpy>=1.20.0",
]
